test(agent): exercise catalog replay selection
