+++ packages/agent/test/peer-event-lifecycle.test.ts
@@ -1,13 +1,10 @@
 import { mkdtemp, rm } from 'node:fs/promises';
 import { tmpdir } from 'node:os';
 import { join } from 'node:path';
-import { PeerEventLifetime } from '../src/p2p/peer-event-lifetime.js';
 import { PeerSyncSession } from '../src/sync/peer-sync-session.js';
-import { syncOpenedPeerConnection } from '../src/sync/peer-connection.js';
 import { describe, expect, it, vi } from 'vitest';
-import { createOperationContext, PROTOCOL_SYNC, PROTOCOL_STORAGE_ACK } from '@origintrail-official/dkg-core';
+import { PROTOCOL_STORAGE_ACK } from '@origintrail-official/dkg-core';
 import { NetworkAdmissionCoordinator } from '../src/p2p/network-admission-coordinator.js';
-import { runSyncOnConnect, runSelectedSharedMemoryRetry, type SyncOnConnectContext } from '../src/sync/on-connect/sync-on-connect.js';
 import { DKGAgent } from '../src/index.js';
 import { MockChainAdapter } from '@origintrail-official/dkg-chain';
 import { asSyncOnConnectTestAgent, allowAllNetworkAdmission } from './_helpers/sync-on-connect-test-fixture.js';
@@ -15,107 +12,7 @@ import { createPeerEventFixture, deferred, flushMicrotasks } from './_helpers/pe
 
 const PROBE = { protocolsKey: null, connectionKey: null } satisfies Awaited<ReturnType<DKGAgent['getSyncReconcilerProbe']>>;
 
-function activeSessionWithoutJobs(): PeerSyncSession {
-  return new PeerSyncSession({
-    createJob: () => { throw new Error('scheduler is outside this fixture'); },
-    onInternalError: () => undefined,
-  });
-}
-
-describe('peer-event lifecycle', () => {
-  it('supplies the fully constructed session when a lazy peer job is created', async () => {
-    let observedSession: PeerSyncSession | undefined;
-    const createJob = vi.fn((_remotePeer: string, session: PeerSyncSession) => {
-      observedSession = session;
-      return {
-        runAutomaticSelectedThenOrdinary: async () => 'not-started' as const,
-        runSelected: async () => 'not-started' as const,
-        cancel: () => undefined,
-        finish: () => undefined,
-      };
-    });
-    const session = new PeerSyncSession({
-      createJob,
-      onInternalError: () => undefined,
-    });
-    try {
-      const scheduler = session.getScheduler();
-      expect(createJob).not.toHaveBeenCalled();
-      expect(scheduler.enqueueOrdinary('peer-a', () => undefined, 0)).toBe(true);
-      await vi.waitFor(() => expect(createJob).toHaveBeenCalledExactlyOnceWith('peer-a', session));
-      expect(observedSession).toBe(session);
-    } finally {
-      session.close();
-    }
-  });
-
-  it('surfaces live supervised errors and prevents work admitted after retirement', async () => {
-    const lifetime = new PeerEventLifetime();
-    const failure = new Error('live task failed');
-    await expect(lifetime.run(async () => { throw failure; })).rejects.toBe(failure);
-    lifetime.close();
-    const release = vi.fn();
-    const unregister = lifetime.onClose(release);
-    expect(release).toHaveBeenCalledOnce();
-    unregister();
-    const work = vi.fn(async () => {});
-    await lifetime.run(work);
-    const commit = vi.fn(() => {});
-    lifetime.commit(commit);
-    expect(commit).not.toHaveBeenCalled();
-    expect(work).not.toHaveBeenCalled();
-  });
-
-  it.each([new Error('queued sync failure'), 'queued sync failure'])('uses active configured replay authority and fences queued errors: %s', async (failure) => {
-    const f = await createPeerEventFixture();
-    const session = activeSessionWithoutJobs();
-    try {
-      vi.spyOn(f.agent, 'listActiveRfc64CatalogReplayContextGraphIdsV1')
-        .mockReturnValue(['active']);
-      const admission = vi.spyOn(f.agent.networkAdmissionCoordinator, 'ensureAdmitted').mockResolvedValue(true);
-      vi.spyOn(f.agent, 'enrichPeerStoreFromInboundCircuit').mockResolvedValue();
-      vi.spyOn(f.agent, 'drainPendingSenderKeyForPeer').mockResolvedValue(0);
-      vi.spyOn(f.agent, 'reannounceRfc64CatalogHeadsToPeerV1').mockResolvedValue({ announced: 0, failed: 0, manifest: [] });
-      const replay = vi.spyOn(f.agent, 'requestRfc64CatalogHeadReplaysFromConnectedPeersV1').mockResolvedValue({ requested: 1, failed: 0 });
-      let reportError!: Parameters<DKGAgent['queueSyncFromPeerOnConnect']>[1];
-      vi.spyOn(f.agent, 'queueSyncFromPeerOnConnect').mockImplementation((_peer, onError) => { reportError = onError; return true; });
-      const log = { info: vi.fn(), warn: vi.fn() };
-      const ports = {
-        localPeerId: f.agent.node.libp2p.peerId.toString(),
-        listActiveCatalogReplayContextGraphIds: () => (
-          f.agent.listActiveRfc64CatalogReplayContextGraphIdsV1()
-        ),
-        markReplayPending: (contextGraphId: string, remotePeer: string) => f.agent.markRfc64CatalogReplayPeerPendingV1(contextGraphId, remotePeer),
-        clearReplayPending: (contextGraphId: string, remotePeer: string) => f.agent.clearRfc64CatalogReplayPeerPendingV1(contextGraphId, remotePeer),
-        ensureAdmitted: (remotePeer: string, ctx: ReturnType<typeof createOperationContext>, signal: AbortSignal) => (
-          f.agent.networkAdmissionCoordinator.ensureAdmitted(remotePeer, ctx, { signal })
-        ),
-        enrichPeerStore: (connection: Parameters<DKGAgent['enrichPeerStoreFromInboundCircuit']>[0]) => (
-          f.agent.enrichPeerStoreFromInboundCircuit(connection)
-        ),
-        drainPendingSenderKey: (remotePeer: string, ctx: ReturnType<typeof createOperationContext>) => (
-          f.agent.drainPendingSenderKeyForPeer(remotePeer, ctx)
-        ),
-        reannounceCatalogHeads: (remotePeer: string) => f.agent.reannounceRfc64CatalogHeadsToPeerV1(remotePeer),
-        requestCatalogReplay: (contextGraphId: string) => f.agent.requestRfc64CatalogHeadReplaysFromConnectedPeersV1(contextGraphId),
-        queueSync: (remotePeer: string, onError: Parameters<DKGAgent['queueSyncFromPeerOnConnect']>[1]) => (
-          f.agent.queueSyncFromPeerOnConnect(remotePeer, onError)
-        ),
-      };
-      const context = { ports, session, ctx: createOperationContext('sync'), log };
-      await syncOpenedPeerConnection(context, { direction: 'inbound', remotePeer: f.agent.node.libp2p.peerId });
-      expect(admission).not.toHaveBeenCalled();
-      await syncOpenedPeerConnection(context, { direction: 'inbound', remotePeer: f.peer });
-      expect(replay).toHaveBeenCalledExactlyOnceWith('active');
-      reportError(f.peerId, failure);
-      expect(log.warn).toHaveBeenCalledExactlyOnceWith(context.ctx, expect.stringContaining('queued sync failure'));
-      log.warn.mockClear();
-      session.close();
-      reportError(f.peerId, failure);
-      expect(log.warn).not.toHaveBeenCalled();
-    } finally { session.close(); await f.close(); }
-  });
-
+describe('DKGAgent peer lifecycle integration', () => {
   it('accepts bootstrap-authorized recovery before listener registration with ordinary sync and retries disabled', async () => {
     const dataDir = await mkdtemp(join(tmpdir(), 'dkg-early-bootstrap-session-'));
     const agent = await DKGAgent.create({
@@ -174,68 +71,6 @@ describe('peer-event lifecycle', () => {
       expect(f.agent.getSyncOnConnectPeerScheduler()).not.toBe(closed);
       expect(f.agent.queueSyncFromPeerOnConnect(f.peerId, () => {}, 60_000)).toBe(true);
     } finally { await f.close(); }
-  });
-
-  it.each([
-    { phase: 'durable', reject: false }, { phase: 'durable', reject: true },
-    { phase: 'discovered', reject: false }, { phase: 'discovered', reject: true },
-    { phase: 'ordinary', reject: false }, { phase: 'ordinary', reject: true },
-    { phase: 'selected', reject: false }, { phase: 'selected', reject: true },
-  ] as const)('fences real $phase transfer continuations (reject=$reject)', async ({ phase, reject }) => {
-    const controller = new AbortController();
-    const gate = deferred<number>();
-    const entered = deferred<void>();
-    const syncingPeers = new Set<string>();
-    const transfer = async () => { entered.resolve(); return gate.promise; };
-    let discovered = false;
-    const sync = vi.fn(async (_peer: string, cgs?: string[]) => {
-      if (phase === 'durable' || (phase === 'discovered' && cgs?.includes('new'))) return transfer();
-      return 1;
-    });
-    const refresh = vi.fn(async () => {});
-    const discover = vi.fn(async () => { discovered = true; return 1; });
-    const ordinary = vi.fn(transfer);
-    const log = vi.fn();
-    const account = vi.fn();
-    const context: SyncOnConnectContext = {
-      signal: controller.signal, remotePeer: 'peer', syncingPeers,
-      getPeerProtocols: async () => [PROTOCOL_SYNC], knownCorePeerIds: new Set(),
-      getSyncContextGraphs: () => phase === 'discovered' && discovered ? ['initial', 'new'] : ['initial'],
-      getDurableSyncContextGraphs: () => ['initial'], syncFromPeer: sync,
-      refreshMetaSyncedFlags: refresh, discoverContextGraphsFromStore: discover,
-      ordinarySharedMemoryLane: { resolveWork: () => ({ contextGraphIds: ['ordinary'], syncFromPeer: ordinary }) },
-      logInfo: log, onSyncAccounting: account,
-    };
-    const attempt = phase === 'selected' ? runSelectedSharedMemoryRetry({
-      signal: controller.signal, remotePeer: 'peer', syncingPeers,
-      getPeerProtocols: context.getPeerProtocols, logInfo: log, onSyncAccounting: account,
-      selectedSharedMemoryLane: { admitWork: () => ({
-        contextGraphIds: ['selected'],
-        syncFromPeer: async () => ({
-          kind: 'selected-shared-memory',
-          requestedScope: { kind: 'selected-public', targets: [{ contextGraphId: 'selected', lane: 'selected-public' }] },
-          scopeComplete: true, selectedScopeComplete: true,
-          targetDiagnostics: { selectedPublic: { completed: 1, total: 1 }, ordinaryPrivate: { completed: 0, total: 0 } },
-          shared: {
-            insertedTriples: await transfer(), fetchedMetaTriples: 0, fetchedDataTriples: 1,
-            insertedMetaTriples: 0, insertedDataTriples: 1, bytesReceived: 1,
-            resumedPhases: 0, timedOutPhases: 0, completedPhases: 1, checkpointAdvances: 0,
-            emptyResponses: 0, droppedDataTriples: 0, failedPeers: 0, failedPhases: 0,
-            deniedPhases: 0, backoffWorthyFailures: 0, deferredBackpressure: 0,
-          },
-        }),
-      }) },
-    }) : runSyncOnConnect(context);
-    const settled = attempt.then(() => undefined, (error: unknown) => error);
-    await entered.promise;
-    expect(syncingPeers.has('peer')).toBe(true);
-    controller.abort();
-    for (const callback of [sync, refresh, discover, ordinary, log, account]) callback.mockClear();
-    if (reject) gate.reject(new Error('retired transfer failed'));
-    else gate.resolve(1);
-    expect(await settled).toBeInstanceOf(Error);
-    for (const callback of [sync, refresh, discover, ordinary, log, account]) expect(callback).not.toHaveBeenCalled();
-    expect(syncingPeers.size).toBe(0);
   });
 
   it.each(['live', 'late success', 'late failure'] as const)('owns already-connected startup admission: %s', async (outcome) => {
@@ -313,271 +148,6 @@ describe('peer-event lifecycle', () => {
       });
     } finally { probeGate.resolve(PROBE); admissionGate.resolve(false); await f.close(); }
   });
-
-  it('keeps active-peer ownership separate when old work settles after a new lifetime starts', async () => {
-    const f = await createPeerEventFixture();
-    const oldGate = deferred<string[]>();
-    const newGate = deferred<string[]>();
-    try {
-      vi.spyOn(f.agent.networkAdmissionCoordinator, 'isAcceptedPeer').mockReturnValue(true);
-      const protocols = vi.spyOn(f.agent, 'getPeerProtocols')
-        .mockReturnValueOnce(oldGate.promise).mockReturnValueOnce(newGate.promise);
-      const oldAttempt = f.agent.trySyncFromPeer(f.peerId).catch(() => undefined);
-      await f.agent.stop();
-      await f.agent.start();
-      vi.spyOn(f.agent.networkAdmissionCoordinator, 'isAcceptedPeer').mockReturnValue(true);
-      const newAttempt = f.agent.trySyncFromPeer(f.peerId);
-      expect(protocols).toHaveBeenCalledTimes(2);
-      oldGate.resolve([]);
-      await oldAttempt;
-      await expect(f.agent.trySyncFromPeer(f.peerId)).resolves.toBe('already-syncing');
-      expect(protocols).toHaveBeenCalledTimes(2);
-      newGate.resolve([]);
-      await expect(newAttempt).resolves.toBe('skipped-no-sync');
-    } finally { oldGate.resolve([]); newGate.resolve([]); await f.close(); }
-  });
-
-  it('cancels queued catch-up synchronously before slow shutdown and permits a new queue after restart', async () => {
-    const f = await createPeerEventFixture();
-    const gate = deferred<void>();
-    try {
-      vi.spyOn(f.agent.networkAdmissionCoordinator, 'isAcceptedPeer').mockReturnValue(true);
-      vi.spyOn(f.agent, 'getSyncReconcilerProbe').mockResolvedValue(PROBE);
-      const attempt = vi.spyOn(f.agent, 'trySyncFromPeer').mockResolvedValue('synced');
-      vi.spyOn(f.agent, 'drainCoreHostRecordings').mockReturnValueOnce(gate.promise);
-      expect(f.agent.queueSyncFromPeerOnConnect(f.peerId, () => {}, 10)).toBe(true);
-      const stopping = f.agent.stop();
-      await new Promise((resolve) => setTimeout(resolve, 30));
-      expect(attempt).not.toHaveBeenCalled();
-      gate.resolve();
-      await stopping;
-      await f.agent.start();
-      vi.spyOn(f.agent.networkAdmissionCoordinator, 'isAcceptedPeer').mockReturnValue(true);
-      expect(f.agent.queueSyncFromPeerOnConnect(f.peerId, () => {}, 0)).toBe(true);
-      await vi.waitFor(() => expect(attempt).toHaveBeenCalledOnce());
-    } finally { gate.resolve(); await f.close(); }
-  });
-
-  it('retires connection-close bookkeeping synchronously before slow shutdown', async () => {
-    const f = await createPeerEventFixture();
-    const gate = deferred<void>();
-    let now = 1_000;
-    try {
-      vi.spyOn(Date, 'now').mockImplementation(() => now);
-      f.dispatchClose();
-      expect(f.state.disconnectTimestamp(f.peerId)).toBe(1_000);
-
-      vi.spyOn(f.agent, 'drainCoreHostRecordings').mockReturnValueOnce(gate.promise);
-      now = 2_000;
-      const stopping = f.agent.stop();
-      expect(f.state.session.isActive()).toBe(false);
-
-      f.dispatchClose();
-      expect(f.state.disconnectTimestamp(f.peerId)).toBe(1_000);
-
-      gate.resolve();
-      await stopping;
-    } finally {
-      gate.resolve();
-      await f.close();
-    }
-  });
-
-  it.each([
-    { entry: 'scheduler', disposition: 'clear' },
-    { entry: 'scheduler', disposition: 'retry' },
-    { entry: 'peer-update', disposition: 'clear' },
-    { entry: 'peer-update', disposition: 'retry' },
-  ] as const)('discards late $entry $disposition accounting across node lifetimes', async ({ entry, disposition }) => {
-    const f = await createPeerEventFixture();
-    const gate = deferred<void>();
-    const entered = deferred<void>();
-    try {
-      vi.spyOn(f.agent.networkAdmissionCoordinator, 'isAcceptedPeer').mockReturnValue(true);
-      vi.spyOn(f.agent, 'getSyncReconcilerProbe').mockResolvedValue(PROBE);
-      const attempt = vi.spyOn(f.agent, 'trySyncFromPeer').mockImplementationOnce(async (_peer, account) => {
-        entered.resolve();
-        await gate.promise;
-        account?.(disposition === 'clear'
-          ? { fresh: true, progress: true, reconcilerDisposition: 'clear' }
-          : { fresh: false, progress: true, reconcilerDisposition: 'retry' });
-        return 'synced';
-      }).mockResolvedValue('synced');
-      vi.spyOn(f.agent, 'isPeerConnectedForSyncBackoff').mockReturnValue(true);
-      if (entry === 'scheduler') f.agent.queueSyncFromPeerOnConnect(f.peerId, () => {}, 0);
-      else { f.state.session.markSkipped(f.peerId); f.dispatchUpdate(); }
-      await entered.promise;
-      await f.agent.stop();
-      await f.agent.start();
-      vi.spyOn(f.agent.networkAdmissionCoordinator, 'isAcceptedPeer').mockReturnValue(true);
-      gate.resolve();
-      await flushMicrotasks();
-      expect(f.state.session.snapshot(f.peerId)).toMatchObject({
-        lastSuccessfulSync: undefined,
-        lastSyncProgress: undefined,
-        backoff: undefined,
-      });
-      expect(f.agent.queueSyncFromPeerOnConnect(f.peerId, () => {}, 0)).toBe(true);
-      await vi.waitFor(() => expect(attempt).toHaveBeenCalledTimes(2));
-    } finally { gate.resolve(); await f.close(); }
-  });
-
-  it('does not mark a peer skipped when an old protocol lookup settles after restart', async () => {
-    const f = await createPeerEventFixture();
-    const gate = deferred<string[]>();
-    try {
-      vi.spyOn(f.agent.networkAdmissionCoordinator, 'isAcceptedPeer').mockReturnValue(true);
-      const protocols = vi.spyOn(f.agent, 'getPeerProtocols').mockReturnValueOnce(gate.promise);
-      const oldAttempt = f.agent.trySyncFromPeer(f.peerId).catch(() => undefined);
-      expect(protocols).toHaveBeenCalledOnce();
-      await f.agent.stop();
-      await f.agent.start();
-      gate.resolve([]);
-      await oldAttempt;
-      expect(f.state.session.snapshot(f.peerId)).toMatchObject({
-        skippedNoSync: false,
-        lastSyncProgress: undefined,
-      });
-    } finally { gate.resolve([]); await f.close(); }
-  });
-
-  it('clears freshness and cooldown state for a peer absent from the stop-time connection list', async () => {
-    const f = await createPeerEventFixture();
-    try {
-      const now = Date.now();
-      f.state.session.recordFreshness(f.peerId, {
-        successfulAt: now,
-        progressAt: now,
-      });
-      f.state.session.recordQueued(f.peerId, now);
-      f.state.session.recordBackoff(
-        f.peerId,
-        { failures: 1, nextRetryAt: now + 60_000, ...PROBE },
-      );
-      expect(f.agent.node.libp2p.getPeers()).not.toContainEqual(f.peer);
-      await f.agent.stop();
-      expect(f.state.session.snapshot(f.peerId)).toMatchObject({
-        lastSuccessfulSync: undefined,
-        lastSyncProgress: undefined,
-        lastQueued: 0,
-        backoff: undefined,
-      });
-      expect(f.state.disconnectTimestamp(f.peerId)).toBeUndefined();
-      await f.agent.start();
-      vi.spyOn(f.agent.networkAdmissionCoordinator, 'isAcceptedPeer').mockReturnValue(true);
-      expect(f.agent.queueSyncFromPeerOnConnect(f.peerId, () => {}, 60_000)).toBe(true);
-    } finally { await f.close(); }
-  });
-
-  it.each(['denied', 'failed'] as const)('clears pending replay when connection admission is %s', async (outcome) => {
-    const f = await createPeerEventFixture();
-    try {
-      vi.spyOn(f.agent, 'readRfc64CatalogResponsibilitiesV1').mockReturnValue([{
-        contextGraphId: 'connection-catalog', responsible: true, active: true, mode: 'catalog',
-        responsibilityReason: 'private-membership', selectionSource: 'default',
-      }]);
-      const admission = vi.spyOn(f.agent.networkAdmissionCoordinator, 'ensureAdmitted');
-      if (outcome === 'denied') admission.mockResolvedValue(false);
-      else admission.mockRejectedValue(new Error('admission unavailable'));
-      const markPending = vi.spyOn(f.agent, 'markRfc64CatalogReplayPeerPendingV1');
-      const clearPending = vi.spyOn(f.agent, 'clearRfc64CatalogReplayPeerPendingV1');
-      const enrich = vi.spyOn(f.agent, 'enrichPeerStoreFromInboundCircuit');
-      const queue = vi.spyOn(f.agent, 'queueSyncFromPeerOnConnect');
-      const warn = vi.spyOn(f.state.log, 'warn').mockImplementation(() => {});
-      f.dispatchOpen();
-      await vi.waitFor(() => expect(clearPending).toHaveBeenCalledWith('connection-catalog', f.peerId));
-      expect(markPending).toHaveBeenCalledWith('connection-catalog', f.peerId);
-      expect(enrich).not.toHaveBeenCalled();
-      expect(queue).not.toHaveBeenCalled();
-      if (outcome === 'failed') expect(warn).toHaveBeenCalledWith(expect.anything(), expect.stringContaining('admission unavailable'));
-    } finally { await f.close(); }
-  });
-
-  it.each(['enrichment', 'sender-key', 'reannouncement', 'replay incomplete'] as const)('continues catch-up after a best-effort %s failure', async (stage) => {
-    const f = await createPeerEventFixture();
-    try {
-      vi.spyOn(f.agent, 'readRfc64CatalogResponsibilitiesV1').mockReturnValue([{
-        contextGraphId: 'connection-catalog', responsible: true, active: true, mode: 'catalog',
-        responsibilityReason: 'private-membership', selectionSource: 'default',
-      }]);
-      vi.spyOn(f.agent.networkAdmissionCoordinator, 'ensureAdmitted').mockResolvedValue(true);
-      const enrich = vi.spyOn(f.agent, 'enrichPeerStoreFromInboundCircuit').mockResolvedValue();
-      const drain = vi.spyOn(f.agent, 'drainPendingSenderKeyForPeer').mockResolvedValue(2);
-      if (stage === 'enrichment') enrich.mockRejectedValue(new Error('enrichment unavailable'));
-      else if (stage === 'sender-key') drain.mockRejectedValue(new Error('sender-key unavailable'));
-      const reannounce = vi.spyOn(f.agent, 'reannounceRfc64CatalogHeadsToPeerV1').mockResolvedValue({ announced: 0, failed: 0, manifest: [] });
-      if (stage === 'reannouncement') reannounce.mockRejectedValue(new Error('reannouncement unavailable'));
-      const replay = vi.spyOn(f.agent, 'requestRfc64CatalogHeadReplaysFromConnectedPeersV1')
-        .mockRejectedValue(new Error('replay unavailable'));
-      if (stage === 'replay incomplete') replay.mockResolvedValue({ requested: 1, failed: 1 });
-      const queue = vi.spyOn(f.agent, 'queueSyncFromPeerOnConnect').mockReturnValue(true);
-      const warn = vi.spyOn(f.state.log, 'warn').mockImplementation(() => {});
-      f.dispatchOpen();
-      await vi.waitFor(() => expect(queue).toHaveBeenCalledWith(f.peerId, expect.any(Function)));
-      await flushMicrotasks();
-      expect(drain).toHaveBeenCalledWith(f.peerId, expect.anything());
-      expect(reannounce).toHaveBeenCalledExactlyOnceWith(f.peerId);
-      expect(replay).toHaveBeenCalledWith('connection-catalog');
-      if (stage === 'replay incomplete') {
-        expect(warn).toHaveBeenCalledWith(expect.anything(), expect.stringContaining('catalog replay incomplete'));
-      } else {
-        expect(warn).toHaveBeenCalledWith(expect.anything(), expect.stringContaining(`${stage} unavailable`));
-        expect(warn).toHaveBeenCalledWith(expect.anything(), expect.stringContaining('replay unavailable'));
-      }
-    } finally { await f.close(); }
-  });
-
-  it.each(['reannouncement rejection', 'replay rejection', 'replay incomplete'] as const)(
-    'ignores old terminal %s after shutdown and restart', async (stage) => {
-      const f = await createPeerEventFixture();
-      const announcementGate = deferred<Awaited<ReturnType<DKGAgent['reannounceRfc64CatalogHeadsToPeerV1']>>>();
-      const replayGate = deferred<Awaited<ReturnType<DKGAgent['requestRfc64CatalogHeadReplaysFromConnectedPeersV1']>>>();
-      try {
-        vi.spyOn(f.agent, 'readRfc64CatalogResponsibilitiesV1').mockReturnValue([{
-          contextGraphId: 'terminal-catalog', responsible: true, active: true, mode: 'catalog',
-          responsibilityReason: 'private-membership', selectionSource: 'default',
-        }]);
-        vi.spyOn(f.agent.networkAdmissionCoordinator, 'ensureAdmitted').mockResolvedValue(true);
-        vi.spyOn(f.agent, 'enrichPeerStoreFromInboundCircuit').mockResolvedValue();
-        vi.spyOn(f.agent, 'drainPendingSenderKeyForPeer').mockResolvedValue(0);
-        const markPending = vi.spyOn(f.agent, 'markRfc64CatalogReplayPeerPendingV1');
-        const clearPending = vi.spyOn(f.agent, 'clearRfc64CatalogReplayPeerPendingV1');
-        const reannounce = vi.spyOn(f.agent, 'reannounceRfc64CatalogHeadsToPeerV1')
-          .mockReturnValue(announcementGate.promise);
-        const replay = vi.spyOn(f.agent, 'requestRfc64CatalogHeadReplaysFromConnectedPeersV1')
-          .mockReturnValue(replayGate.promise);
-        const queue = vi.spyOn(f.agent, 'queueSyncFromPeerOnConnect').mockReturnValue(true);
-        const warn = vi.spyOn(f.state.log, 'warn').mockImplementation(() => {});
-        f.dispatchOpen();
-        await vi.waitFor(() => expect(queue).toHaveBeenCalledOnce());
-        expect(reannounce).toHaveBeenCalledExactlyOnceWith(f.peerId);
-        expect(replay).toHaveBeenCalledExactlyOnceWith('terminal-catalog');
-        expect(markPending).toHaveBeenCalledExactlyOnceWith('terminal-catalog', f.peerId);
-        await f.agent.stop();
-        expect(clearPending).toHaveBeenCalledExactlyOnceWith('terminal-catalog', f.peerId);
-        await f.agent.start();
-        // A new lifetime may already have acquired another pending-peer fence.
-        f.agent.markRfc64CatalogReplayPeerPendingV1('terminal-catalog', f.peerId);
-        clearPending.mockClear();
-        warn.mockClear();
-        queue.mockClear();
-        if (stage === 'reannouncement rejection') announcementGate.reject(new Error('old announcement'));
-        else announcementGate.resolve({ announced: 1, failed: 0, manifest: [] });
-        if (stage === 'replay rejection') replayGate.reject(new Error('old replay'));
-        else replayGate.resolve({ requested: 1, failed: 1 });
-        await flushMicrotasks();
-        expect(warn).not.toHaveBeenCalled();
-        expect(queue).not.toHaveBeenCalled();
-        expect(clearPending).not.toHaveBeenCalled();
-        expect(replay).toHaveBeenCalledOnce();
-        expect(reannounce).toHaveBeenCalledOnce();
-      } finally {
-        announcementGate.resolve({ announced: 0, failed: 0, manifest: [] });
-        replayGate.resolve({ requested: 0, failed: 0 });
-        await f.close();
-      }
-    },
-  );
 
   it('records the offline boundary so a same-instance restart immediately queues catch-up', async () => {
     const f = await createPeerEventFixture();
@@ -661,76 +231,6 @@ describe('peer-event lifecycle', () => {
     } finally {
       admissionGate.resolve(false);
       probeGate.resolve(PROBE);
-      await f.close();
-    }
-  });
-
-  it.each([
-    { name: 'event', gate: 'event', reject: false },
-    { name: 'admission', gate: 'admission', reject: false },
-    { name: 'admission abort', gate: 'admission', reject: true },
-    { name: 'enrichment', gate: 'enrichment', reject: false },
-    { name: 'sender-key', gate: 'sender-key', reject: false },
-  ] as const)('does not resume connection work after shutdown at $name', async (scenario) => {
-    const f = await createPeerEventFixture();
-    const admissionGate = deferred<void>();
-    const enrichmentGate = deferred<void>();
-    const senderKeyGate = deferred<void>();
-    try {
-      vi.spyOn(f.agent, 'readRfc64CatalogResponsibilitiesV1').mockReturnValue([{
-        contextGraphId: 'shutdown-catalog', responsible: true, active: true, mode: 'catalog',
-        responsibilityReason: 'private-membership', selectionSource: 'default',
-      }]);
-      const markPending = vi.spyOn(f.agent, 'markRfc64CatalogReplayPeerPendingV1');
-      const clearPending = vi.spyOn(f.agent, 'clearRfc64CatalogReplayPeerPendingV1');
-      const admission = vi.spyOn(f.agent.networkAdmissionCoordinator, 'ensureAdmitted').mockImplementation(async (_peer, _ctx, options) => {
-        if (scenario.reject) {
-          options?.signal?.addEventListener('abort', () => admissionGate.reject(new Error('admission cancelled')), { once: true });
-        }
-        if (scenario.gate === 'admission') await admissionGate.promise;
-        return true;
-      });
-      const enrich = vi.spyOn(f.agent, 'enrichPeerStoreFromInboundCircuit').mockImplementation(async () => {
-        if (scenario.gate === 'enrichment') await enrichmentGate.promise;
-      });
-      const drain = vi.spyOn(f.agent, 'drainPendingSenderKeyForPeer').mockImplementation(async () => {
-        if (scenario.gate === 'sender-key') await senderKeyGate.promise;
-        return 0;
-      });
-      const queue = vi.spyOn(f.agent, 'queueSyncFromPeerOnConnect').mockReturnValue(false);
-      const reannounce = vi.spyOn(f.agent, 'reannounceRfc64CatalogHeadsToPeerV1').mockResolvedValue({ announced: 0, failed: 0, manifest: [] });
-      const replay = vi.spyOn(f.agent, 'requestRfc64CatalogHeadReplaysFromConnectedPeersV1').mockResolvedValue({ requested: 0, failed: 0 });
-      let stopping: Promise<void>;
-      if (scenario.gate === 'event') {
-        stopping = f.agent.stop();
-        f.dispatchOpen();
-      } else {
-        f.dispatchOpen();
-        await flushMicrotasks();
-        expect(admission).toHaveBeenCalledOnce();
-        if (scenario.gate === 'enrichment') expect(enrich).toHaveBeenCalledOnce();
-        if (scenario.gate === 'sender-key') expect(drain).toHaveBeenCalledOnce();
-        stopping = f.agent.stop();
-      }
-      await stopping;
-      if (scenario.gate !== 'event') {
-        expect(markPending).toHaveBeenCalledExactlyOnceWith('shutdown-catalog', f.peerId);
-        expect(clearPending).toHaveBeenCalledExactlyOnceWith('shutdown-catalog', f.peerId);
-      }
-      admissionGate.resolve();
-      enrichmentGate.resolve();
-      senderKeyGate.resolve();
-      await flushMicrotasks();
-      if (scenario.gate === 'event') expect(admission).not.toHaveBeenCalled();
-      if (scenario.gate === 'event' || scenario.gate === 'admission') expect(enrich).not.toHaveBeenCalled();
-      if (scenario.gate !== 'sender-key') expect(drain).not.toHaveBeenCalled();
-      expect(reannounce).not.toHaveBeenCalled();
-      expect(replay).not.toHaveBeenCalled();
-      expect(queue).not.toHaveBeenCalled();
-    } finally {
-      admissionGate.resolve();
-      enrichmentGate.resolve();
-      senderKeyGate.resolve();
       await f.close();
     }
   });
